RDKBDEV-454: fix misuse of "\0" & drop legacy deps
Reason for change:

fix misuse of "\0"

"\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).

Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

read from /version.txt instead of /fss/gw/version.txt

Access the actual version.txt file directly rather than relying on
the /fss/gw/version.txt symlink.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

drop dependency on legacy /fss/gw symlinks

Access /usr/ccsp/tr069pa and /rdklogger/backupLogs.sh directly rather
than via the legacy symlinks in /fss/gw

Test Procedure:
 - get sw version param thru tr69
 - basic sanity + tr69
 - set param & get param using tr69
 - reboot from acs (logs should be tarred and uploaded to acs server)

Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I7f531e66c3de40d1f6161f687e9b2757dbefa2aa