RDKCOM-2289: RDKBDEV-453: add -lcjson & -lprivilege
Reason for change:
add -lcjson to CcspTr069PaSsp_LDFLAGS
The code needs to link with libcjson, so update the Makefiles to do
so rather than relying on the recipe to provide a workaround.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
add -lprivilege to CcspTr069PaSsp_LDFLAGS
The code needs to link with libprivilege, so update the Makefiles to
do so rather than relying on the recipe to provide a workaround.
Test Procedure:
1. Yocto build should pass
2. Tr069 should be up and running with Basic Sanity
Risks: None.
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

(cherry picked from commit 5ca181e2500474dadbb5b97d39209c32e94a5e8b)
Change-Id: Ide19357e8617021979b0d6a2aeb354927e459ef3