RDKBDEV-458: minor code cleanup
Reason for change:

minor code cleanup in CcspCwmppoGetAcsInfo()

Use sizeof(pProperty->AcsUrl) as the size limit for the new URL
instead of assuming the buffer size equals CCSP_CWMP_MAX_URL_SIZE (it
does so now but it's more robust to use sizeof() than to hardcode).

If the new URL is too long then ignore it. The previous approach was
to truncate but that just results in an invalid URL.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

minor code cleanup in CcspCwmppoSetupEnv()

Use sizeof(pProperty->AcsUrl) as the size limit for the new URL
instead of assuming the buffer size equals CCSP_CWMP_MAX_URL_SIZE (it
does so now but it's more robust to use sizeof() than to hardcode).

Test Procedure:
1. Check the process should be up and running.
2. Tr069 Sanity test shoud be sufficeint.
3. Check the tr069 process should not crash or restart during test.

Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I282e0de377596cd58fffbdcfcb58f179e221478c