RDKBDEV-1324: make intHttpContent variable local
Reason for change:
make intHttpContent variable local to CcspCwmpAcscoRequest() instead of global
Unprotected global variables are not safe in code which may be used
in a multi-threaded environment. Since there's no clear reason for
intHttpContent to be a global variable, fix by making it a local
variable within CcspCwmpAcscoRequest(). Also drop duplicate zeroing
of the contents of the variable.

Test Procedure: Sanity.
Risks: None.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>
Change-Id: I2c43712886a29e8aa20732a7e25ade19074a786e